Move mixin for CachedTor into CachedBitcoinSAppConfigCachedTor so subclasses dont need to do it
