+++ testkit/src/main/scala/org/bitcoins/testkit/rpc/BitcoindRpcTorTest.scala
@@ -1,16 +1,13 @@
 package org.bitcoins.testkit.rpc
 
 import org.bitcoins.testkit.node.{CachedBitcoinSAppConfigCachedTor}
-import org.bitcoins.testkit.tor.CachedTor
 import org.bitcoins.testkit.util.BitcoindRpcTest
 
 trait BitcoindRpcTorTest
     extends BitcoindRpcTest
-    with CachedBitcoinSAppConfigCachedTor
-    with CachedTor {
+    with CachedBitcoinSAppConfigCachedTor {
 
   override def beforeAll(): Unit = {
-    super[CachedTor].beforeAll()
     super[CachedBitcoinSAppConfigCachedTor].beforeAll()
     super[BitcoindRpcTest].beforeAll()
   }
@@ -18,6 +15,5 @@ trait BitcoindRpcTorTest
   override def afterAll(): Unit = {
     super[BitcoindRpcTest].afterAll()
     super[CachedBitcoinSAppConfigCachedTor].afterAll()
-    super[CachedTor].afterAll()
   }
 }
